+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# v1.0
+# v1.0.2
+## 21/01/2019
+
+1. [Change](#Change)
+    * Add comment to Cadphp class
+	* Remove Variables Type for PHP 5.6.4 compatibily Issue #1
+	
+
+# v1.0.1
 ## 23/12/2018
 
 1. [New](#new)